Allow address resolver override

+++ include/RED4ext/Relocation-inl.hpp
@@ -10,6 +10,7 @@
 #include <Windows.h>
 
 #include <RED4ext/Common.hpp>
+#include <RED4ext/Detail/Memory.hpp>
 
 RED4EXT_INLINE uintptr_t RED4ext::RelocBase::GetImageBase()
 {
@@ -20,48 +21,55 @@ RED4EXT_INLINE uintptr_t RED4ext::RelocBase::GetImageBase()
 RED4EXT_INLINE
 uintptr_t RED4ext::UniversalRelocBase::Resolve(uint32_t aHash)
 {
-    using functionType = uintptr_t (*)(uint32_t);
-    static functionType resolveFunc = nullptr;
-
-    static std::once_flag flag;
-    std::call_once(flag,
-       []()
-       {
-           constexpr auto dllName = "RED4ext.dll";
-           constexpr auto functionName = "RED4ext_ResolveAddress";
-
-           auto handle = GetModuleHandleA(dllName);
-           if (!handle)
-           {
-               std::stringstream stream;
-               stream << "Failed to get '" << dllName
-                      << "' handle.\nProcess will now close.\n\nLast error: " << GetLastError();
-
-               MessageBoxA(nullptr, stream.str().c_str(), "RED4ext.SDK", MB_ICONERROR | MB_OK);
-               TerminateProcess(GetCurrentProcess(), 1);
-           }
-
-           resolveFunc = reinterpret_cast<functionType>(GetProcAddress(handle, functionName));
-           if (resolveFunc == nullptr)
-           {
-               std::stringstream stream;
-               stream << "Failed to get '" << functionName
-                      << "' address.\nProcess will now close.\n\nLast error: " << GetLastError();
-
-               MessageBoxA(nullptr, stream.str().c_str(), "RED4ext.SDK", MB_ICONERROR | MB_OK);
-               TerminateProcess(GetCurrentProcess(), 1);
-           }
-       });
-
-    auto address = resolveFunc(aHash);
-    if (address == 0)
+    if constexpr (Detail::AddressResolverOverride<uint32_t>::value)
     {
-        std::stringstream stream;
-        stream << "Failed to resolve address for hash " << std::hex << std::showbase << aHash << ".\nProcess will now close.";
-
-        MessageBoxA(nullptr, stream.str().c_str(), "RED4ext.SDK", MB_ICONERROR | MB_OK);
-        TerminateProcess(GetCurrentProcess(), 1);
+        return Detail::AddressResolverOverride<uint32_t>::Resolve(aHash);
     }
+    else
+    {
+        using functionType = uintptr_t (*)(uint32_t);
+        static functionType resolveFunc = nullptr;
 
-    return address;
+        static std::once_flag flag;
+        std::call_once(flag,
+           []()
+           {
+               constexpr auto dllName = "RED4ext.dll";
+               constexpr auto functionName = "RED4ext_ResolveAddress";
+
+               auto handle = GetModuleHandleA(dllName);
+               if (!handle)
+               {
+                   std::stringstream stream;
+                   stream << "Failed to get '" << dllName
+                          << "' handle.\nProcess will now close.\n\nLast error: " << GetLastError();
+
+                   MessageBoxA(nullptr, stream.str().c_str(), "RED4ext.SDK", MB_ICONERROR | MB_OK);
+                   TerminateProcess(GetCurrentProcess(), 1);
+               }
+
+               resolveFunc = reinterpret_cast<functionType>(GetProcAddress(handle, functionName));
+               if (resolveFunc == nullptr)
+               {
+                   std::stringstream stream;
+                   stream << "Failed to get '" << functionName
+                          << "' address.\nProcess will now close.\n\nLast error: " << GetLastError();
+
+                   MessageBoxA(nullptr, stream.str().c_str(), "RED4ext.SDK", MB_ICONERROR | MB_OK);
+                   TerminateProcess(GetCurrentProcess(), 1);
+               }
+           });
+
+        auto address = resolveFunc(aHash);
+        if (address == 0)
+        {
+            std::stringstream stream;
+            stream << "Failed to resolve address for hash " << std::hex << std::showbase << aHash << ".\nProcess will now close.";
+
+            MessageBoxA(nullptr, stream.str().c_str(), "RED4ext.SDK", MB_ICONERROR | MB_OK);
+            TerminateProcess(GetCurrentProcess(), 1);
+        }
+
+        return address;
+    }
 }
